# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BundleSummaryResponse": {
        "properties": {
          "contentHash": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "sequence": {
            "format": "int64",
            "type": "integer"
          },
          "siteCount": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "CertificateResponse": {
        "description": "One certificate, without any key material.",
        "properties": {
          "autoRenew": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "daysRemaining": {
            "format": "int64",
            "type": "integer"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "fingerprintSha256": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "inUse": {
            "type": "boolean"
          },
          "issuer": {
            "type": "string"
          },
          "lastError": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "notBefore": {
            "format": "date-time",
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "renewBeforeDays": {
            "format": "int32",
            "type": "integer"
          },
          "revokedAt": {
            "format": "date-time",
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "autoRenew",
          "createdAt",
          "createdBy",
          "domains",
          "id",
          "inUse",
          "name",
          "provider",
          "renewBeforeDays",
          "status",
          "version"
        ],
        "type": "object"
      },
...
      "RequestAcmeRequest": {
        "properties": {
          "autoRenew": {
            "type": "boolean"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array"
          },
          "name": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "renewBeforeDays": {
            "format": "int32",
            "maximum": 89,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "domains",
          "name"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Management API for reverse-proxy sites across one or more NGINX hosts.\n\nAuthentication is OIDC: obtain an access token from Keycloak and send it as\na bearer token. Errors are RFC 9457 problem documents; branch on the stable\n`type` URI rather than on the human-readable `detail`.\n",
    "license": {
      "name": "Proprietary"
    },
    "title": "Easy NGINX Admin API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/certificates": {
      "get": {
        "description": "Metadata only. Private key material is never returned by any endpoint.",
        "operationId": "list_4",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CertificateResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List certificates",
        "tags": [
          "Certificates"
        ]
      },
      "post": {
        "description": "Validates with HTTP-01 against the managed NGINX hosts. Wildcards need DNS-01 and are refused. Requires MANAGE over the namespace of every domain requested, because a certificate is authority to terminate TLS for that name.",
        "operationId": "request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAcmeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Request a certificate over ACME",
        "tags": [
          "Certificates"
        ]
      }
    },
...
    "/api/v1/nginx-instances/{id}/bundles": {
      "get": {
        "operationId": "bundles",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BundleSummaryResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Recent configuration bundles for an instance, newest first",
        "tags": [
          "Deployments"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer-jwt": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
...
}